feat: MTP v3 training pipeline — KL distillation, step-based stages, GWMT export
Training redesign from v2 (hard targets, CE loss) to v3 (soft targets, KL
divergence with T=2.0). Key changes:
- Step-based stage transition (--stage1-steps 1000) instead of epoch-based
- Stage 1: lm_head only at lr=1e-3, Stage 2: all params at lr=1e-4
- Resume with auto stage detection from checkpoint global_step
- Hidden noise injection U(-0.05, 0.05) for regularization (EAGLE)
- Early stopping on val_loss with patience=3
- GWMT v3 export with embedded lm_head + restricted vocab footer
- FP32 training with autocast (was FP16, broke GradScaler)
- Fixed tok/s display on resume (was inflated by carried-over token count)

Results: 28% → 71.2% acceptance at K=4096, early stopped at step 22049.

Tests: acceptance rate validated against pre-trained baseline on eval set
Perf: 71.2% acceptance (K=4096), val_loss=0.5216

+++ train/model.py
@@ -350,8 +350,8 @@ class MTPHead(nn.Module):
             print(f"Loaded {loaded}/15 pre-trained MTP tensors (lm_head initialized randomly)")
 
         model.load_state_dict(state_dict, strict=False)
-        # Convert to FP16 to match CUDA inference precision (GWMT export does BF16→FP16)
-        model = model.half()
+        # Keep FP32 for training (GradScaler needs FP32 gradients).
+        # autocast handles FP16 forward pass. Export converts to FP16.
         return model.to(device)
 
     def param_count(self) -> dict[str, int]:
